Allow caching regular expression matching in rules
This drastically improves performance for some use cases.

The exporter is configured with a set of rules to match regular
expressions against bean names. Regular expressions are CPU-intensive,
and matching the same bean names over the same patterns is wasteful when
there are tens of thousands of bean names (e.g: exposed by Kafka).
Using a single `ConcurrentHashMap` to store the result of pattern
matching here.

The `CachedRule` class was introduced here to "store" this result.
Each rule in the configuration leads to a `CachedRule` which helps
build the set of metrics to expose to Prometheus.
The mapping between a bean name + attributes to a `CachedRule` is done
once. Some (in Kafka, most) bean names may not match any rule, so
caching of `CachedRule.invalid()` helps prevent re-computation of
bean names to non-existent rules.

The logic to map a bean name to a Prometheus MetricFamilySample is
unchanged from the upstream, the CachedRule was introduced only to
be able to cache the results.

Caching is disabled by default (behaviour unchanged), can be enabled by
setting `cacheRules: true` in the configuration.