Allow caching MBeanInfo attribute fetching
Each MBean has a list of attributes (e.g:
kafka.server:type=BrokerTopicMetrics,name=TotalFetchRequestsPerSec,topic=<topic>)
has Count, MeanRate, FiveMinRate, etc...
These do not change over time for most applications and can be cached.
Knowing that some applications (e.g Kafka) can have tens of thousands
of these (per-topic metrics), it saves a lot of calls to the mbean server.

Changes to this are part of the `JmxMBeanPropertyCache` class, where
`getAttributes` was added. It uses a simple ConcurrentHashMap to store
attributes of a bean. Prior art is present in the Cassandra Exporter [1].

This changes the flow of calls to the MBean server a little, instead of
doing: getMBeanInfo(mbeanName) -> getAttributes(mbeanName, attributes)
The flow is now: getMBeanInfo(mbeanName).getAttributes() -> foreach(attribute) getAttribute(mbeanName, attribute).

The main difference is that getAttributes() can be cached, while
getAttributes(mbeanName, attributes) cannot be cached as it also fetches
the value of the attribute [2].

While most applications have immutable MBean attributes, this is not a
strict requirement from the JVM. Applications can opt-out of using the
cache by using `cacheMBeanAttributeInfo: false` in the configuration
(disabled by default to keep backwards compatibility).

[1] https://github.com/criteo/cassandra_exporter/blob/03db26a973f41f02d2cc2d1eae0e6fe090637e08/src/main/java/com/criteo/nosql/cassandra/exporter/JmxScraper.java#L232
[2] https://docs.oracle.com/javase/8/docs/api/javax/management/MBeanServerConnection.html#getAttributes-javax.management.ObjectName-java.lang.String:A-